+++ mainapp/src/Component/SourcesTable.tsx
@@ -47,7 +47,6 @@ import * as z from "zod";
 
 const SourcesTable = () => {
     const { model, updateModel } = useModel();
-
     const [filteringChars, setFilteringChars] = React.useState("");
     const [orderBy, setOrderBy] = React.useState<keyof ServerSource>("name");
     const [order, setOrder] = React.useState<Order>("asc");
@@ -57,6 +56,14 @@ const SourcesTable = () => {
         setOrder(isAsc ? "desc" : "asc");
         setOrderBy(property);
     }
+    const [me, setMe] = React.useState("");
+    React.useEffect(() => {
+        (async () => {
+            const response = await fetch("/api/v1/auth/whoami");
+            const json = (await response.json()) as Response;
+            setMe(json.user.login);
+        })();
+    }, []);
 
     const indices = SRD.withDefault(null, model.indices);
     const graphs = SRD.withDefault(null, model.graphs);
@@ -114,7 +121,7 @@ const SourcesTable = () => {
                             }}
                             renderInput={(params) => <TextField {...params} label="Search Sources by name" />}
                         />
-                        <TableContainer sx={{ maxHeight: "400px" }} component={Paper}>
+                        <TableContainer sx={{ height: "400px" }} component={Paper}>
                             <Table stickyHeader>
                                 <TableHead>
                                     <TableRow>
@@ -147,7 +154,6 @@ const SourcesTable = () => {
                                         .map((source) => {
                                             const haveIndices = indices ? indices.includes(source.name.toLowerCase()) : false;
                                             const haveGraphs = graphs ? graphs.includes(source.graphUri || "") : false;
-                                            console.log();
                                             return (
                                                 <TableRow key={source.name}>
                                                     <TableCell>{source.name}</TableCell>
@@ -167,7 +173,7 @@ const SourcesTable = () => {
                                                     </TableCell>
                                                     <TableCell align="center">
                                                         <Stack direction="row" justifyContent="center" spacing={{ xs: 1 }} useFlexGap>
-                                                            <SourceForm source={source} />
+                                                            <SourceForm source={source} me={me} />
                                                             <ButtonWithConfirmation label="Delete" msg={() => deleteSource(source, updateModel)} />
                                                         </Stack>
                                                     </TableCell>
@@ -181,7 +187,7 @@ const SourcesTable = () => {
                             <CsvDownloader separator="&#9;" filename="sources" extension=".tsv" datas={datas as Datas}>
                                 <Button variant="outlined">Download CSV</Button>
                             </CsvDownloader>
-                            <SourceForm create={true} />
+                            <SourceForm create={true} me={me} />
                         </Stack>
                     </Stack>
                 );
@@ -302,7 +308,7 @@ type SourceFormProps = {
     create?: boolean;
 };
 
-const SourceForm = ({ source = defaultSource(ulid()), create = false }: SourceFormProps) => {
+const SourceForm = ({ source = defaultSource(ulid()), create = false, me = "" }: SourceFormProps) => {
     const { model, updateModel } = useModel();
     const unwrappedSources = SRD.unwrap([], identity, model.sources);
     const sources = React.useMemo(() => unwrappedSources, [unwrappedSources]);
@@ -329,6 +335,10 @@ const SourceForm = ({ source = defaultSource(ulid()), create = false }: SourceFo
     };
 
     const _handleFieldUpdate = (event: React.ChangeEvent<HTMLInputElement>) => update({ type: Type.UserAddedGraphUri, payload: event.target.value });
+
+    if (sourceModel.sourceForm.owner.length == 0) {
+        update({ type: Type.UserUpdatedField, payload: { fieldname: "owner", newValue: me } });
+    }
 
     const handleSparql_serverUpdate = (fieldName: string) => (event: React.ChangeEvent<HTMLTextAreaElement>) => {
         update({
@@ -423,7 +433,7 @@ const SourceForm = ({ source = defaultSource(ulid()), create = false }: SourceFo
                         <Grid item>
                             <Grid alignItems="center" container wrap="nowrap">
                                 <Grid item flex={1}>
-                                    <FormControlLabel control={<Checkbox checked={sourceModel.sourceForm.editable} onChange={handleCheckbox("editable")} />} label="Is this source editable?" />
+                                    <FormControlLabel control={<Checkbox checked={sourceModel.sourceForm.editable} onChange={handleCheckbox("editable")} />} label="Editable?" />
                                 </Grid>
                                 <Grid item>
                                     <HelpButton title="Editable" message={sourceHelp.editable} />
@@ -433,7 +443,7 @@ const SourceForm = ({ source = defaultSource(ulid()), create = false }: SourceFo
                         <Grid item>
                             <Grid alignItems="center" container wrap="nowrap">
                                 <Grid item flex={1}>
-                                    <FormControlLabel control={<Checkbox checked={sourceModel.sourceForm.isDraft} onChange={handleCheckbox("isDraft")} />} label="Is it a draft?" />
+                                    <FormControlLabel control={<Checkbox checked={sourceModel.sourceForm.isDraft} onChange={handleCheckbox("isDraft")} />} label="Draft?" />
                                 </Grid>
                                 <Grid item>
                                     <HelpButton title="isDraft" message={sourceHelp.isDraft} />
@@ -450,6 +460,16 @@ const SourceForm = ({ source = defaultSource(ulid()), create = false }: SourceFo
                                 </Grid>
                                 <Grid item>
                                     <HelpButton title="allowIndividuals" message={sourceHelp.allowIndividuals} />
+                                </Grid>
+                            </Grid>
+                        </Grid>
+                        <Grid item>
+                            <Grid alignItems="center" container wrap="nowrap">
+                                <Grid item flex={1}>
+                                    <FormControlLabel control={<Checkbox checked={sourceModel.sourceForm.published} onChange={handleCheckbox("published")} />} label="Published?" />
+                                </Grid>
+                                <Grid item>
+                                    <HelpButton title="published" message={sourceHelp.published} />
                                 </Grid>
                             </Grid>
                         </Grid>
@@ -631,24 +651,50 @@ const SourceForm = ({ source = defaultSource(ulid()), create = false }: SourceFo
                             </FormControl>
                         </Grid>
                         <Grid item xs={6}>
+                            <FormControl>
+                                <InputLabel id="owner">owner</InputLabel>
+                                <Select
+                                    labelId="owner"
+                                    id="owner-select"
+                                    value={sourceModel.sourceForm.owner}
+                                    label="select-owner"
+                                    fullWidth
+                                    style={{ width: "400px" }}
+                                    renderValue={(selected: string | string[]) => (typeof selected === "string" ? selected : selected.join(", "))}
+                                    onChange={handleFieldUpdate("owner")}
+                                    endAdornment={
+                                        <InputAdornment position="end">
+                                            <HelpButton title="owner" message={sourceHelp.owner} />
+                                        </InputAdornment>
+                                    }
+                                >
+                                    {model.users.data.map((user) => (
+                                        <MenuItem key={user.id} value={user.login}>
+                                            {user.login}
+                                        </MenuItem>
+                                    ))}
+                                </Select>
+                            </FormControl>
+                        </Grid>
+                        <Grid item xs={6}>
                             <Autocomplete
                                 freeSolo
                                 disableClearable
                                 options={knownGroup}
                                 label={"Group"}
+                                onInputChange={(_e, newValue) => handleGroupUpdate(newValue)}
+                                onBlur={validateAfterSubmission}
+                                inputValue={sourceModel.sourceForm.group}
+                                id="group"
                                 renderInput={(params) => (
                                     <TextField
                                         {...params}
                                         helperText={errorMessage(zo.errors.group)}
-                                        id="group"
                                         label="Group"
-                                        name={zo.fields.group()}
-                                        onBlur={validateAfterSubmission}
-                                        onChange={handleFieldUpdate("group")}
-                                        value={sourceModel.sourceForm.group}
                                         InputProps={{
                                             ...params.InputProps,
                                             type: "search",
+                                            name: zo.fields.group(),
                                         }}
                                     />
                                 )}